feat: moved get_next_path function into utils.py
